Move DataManager to data package
Minor cleanups
Full Support for MySQL stuffs
