+++ src/com/Jessy1237/DwarfCraft/DwarfCraft.java
@@ -26,6 +26,7 @@ import org.bukkit.plugin.RegisteredServiceProvider;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import com.Jessy1237.DwarfCraft.commands.*;
+import com.Jessy1237.DwarfCraft.data.DataManager;
 import com.Jessy1237.DwarfCraft.listeners.DwarfBlockListener;
 import com.Jessy1237.DwarfCraft.listeners.DwarfEntityListener;
 import com.Jessy1237.DwarfCraft.listeners.DwarfInventoryListener;
