+++ src/com/Jessy1237/DwarfCraft/data/DBWrapperMySQL.java
@@ -1,14 +1,4 @@
-/*
- * Copyright (c) 2018.
- *
- * DwarfCraft is an RPG plugin that allows players to improve their characters
- * skills and capabilities through training, not experience.
- *
- * Authors: Jessy1237 and Drekryan
- * Original Authors: smartaleq, LexManos and RCarretta
- */
-
-package com.Jessy1237.DwarfCraft;
+package com.Jessy1237.DwarfCraft.data;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -20,51 +10,58 @@ import java.util.HashMap;
 import java.util.UUID;
 import java.util.logging.Level;
 
-import org.bukkit.Location;
-import org.bukkit.entity.EntityType;
-
+import com.Jessy1237.DwarfCraft.ConfigManager;
+import com.Jessy1237.DwarfCraft.DwarfCraft;
 import com.Jessy1237.DwarfCraft.models.DwarfPlayer;
 import com.Jessy1237.DwarfCraft.models.DwarfSkill;
-import com.Jessy1237.DwarfCraft.models.DwarfTrainerTrait;
 
-import net.citizensnpcs.api.npc.AbstractNPC;
-
-class DBWrapperSQLite
+class DBWrapperMySQL
 {
     private final ConfigManager configManager;
     private final DwarfCraft plugin;
     private Connection mDBCon;
 
-    DBWrapperSQLite( DwarfCraft plugin, ConfigManager cm )
+    private String host;
+    private int port;
+    private String database;
+    private String username;
+    private String password;
+
+    DBWrapperMySQL( DwarfCraft plugin, ConfigManager cm )
     {
         this.plugin = plugin;
         this.configManager = cm;
+
+        this.host = configManager.host;
+        this.port = configManager.port;
+        this.database = configManager.database;
+        this.username = configManager.username;
+        this.password = configManager.password;
     }
 
     @SuppressWarnings( "deprecation" )
-    void dbInitialize()
+    protected void dbInitialize()
     {
         try
         {
-            Class.forName( "org.sqlite.JDBC" );
-            mDBCon = DriverManager.getConnection( "jdbc:sqlite:" + configManager.getDbPath() );
+            Class.forName( "com.mysql.jdbc.Driver" );
+            mDBCon = DriverManager.getConnection( "jdbc:mysql://" + host + ":" + port + "/" + database, username, password );
             Statement statement = mDBCon.createStatement();
-            ResultSet rs = statement.executeQuery( "select * from sqlite_master WHERE name = 'players';" );
-            if ( !rs.next() )
-            {
-                buildDB();
-            }
+            ResultSet rs = null;
 
-            // check for update to skill deposits
+            // Ensure database is valid
+            buildDB();
+
+            // Ensure skill deposits exist
             try
             {
                 rs = statement.executeQuery( "SELECT deposit1 FROM skills;" );
             }
             catch ( SQLException ex )
             {
-                statement.executeUpdate( "ALTER TABLE skills ADD COLUMN deposit1 NUMERIC DEFAULT 0;" );
-                statement.executeUpdate( "ALTER TABLE skills ADD COLUMN deposit2 NUMERIC DEFAULT 0;" );
-                statement.executeUpdate( "ALTER TABLE skills ADD COLUMN deposit3 NUMERIC DEFAULT 0;" );
+                statement.executeUpdate( "ALTER TABLE skills ADD COLUMN deposit1 INT DEFAULT 0;" );
+                statement.executeUpdate( "ALTER TABLE skills ADD COLUMN deposit2 INT DEFAULT 0;" );
+                statement.executeUpdate( "ALTER TABLE skills ADD COLUMN deposit3 INT DEFAULT 0;" );
             }
 
             // Adds the uuid arg to the player table
@@ -74,7 +71,7 @@ class DBWrapperSQLite
             }
             catch ( Exception e )
             {
-                plugin.getLogger().log( Level.INFO, "Converting Player DB (may lag a little wait for completion message)." );
+                plugin.getLogger().log( Level.INFO, "Converting Player Database..." );
                 mDBCon.setAutoCommit( false );
                 HashMap<UUID, String> dcplayers = new HashMap<>();
                 HashMap<UUID, Integer> ids = new HashMap<>();
@@ -96,12 +93,12 @@ class DBWrapperSQLite
                     e1.printStackTrace();
                 }
                 statement.executeUpdate( "DROP TABLE players" );
-                statement.executeUpdate( "create table players ( id INTEGER PRIMARY KEY, uuid, race, raceMaster );" );
+                statement.executeUpdate( "CREATE TABLE IF NOT EXISTS `" + database + "`. `players` ( `id` INT NOT NULL AUTO_INCREMENT, `uuid` TEXT, `race` TEXT, `raceMaster` TEXT, PRIMARY KEY( `id` ))ENGINE = InnoDB;" );
                 for ( UUID uuid : dcplayers.keySet() )
                 {
                     if ( uuid != null )
                     {
-                        PreparedStatement prep = mDBCon.prepareStatement( "insert into players(id, uuid, race, raceMaster) values(?,?,?,?);" );
+                        PreparedStatement prep = mDBCon.prepareStatement( "INSERT INTO players(id, uuid, race, raceMaster) values(?,?,?,?);" );
                         prep.setInt( 1, ids.get( uuid ) );
                         prep.setString( 2, uuid.toString() );
                         prep.setString( 3, dcplayers.get( uuid ) );
@@ -110,7 +107,7 @@ class DBWrapperSQLite
                         prep.close();
                     }
                 }
-                plugin.getLogger().log( Level.INFO, "Finished Converting the Players DB." );
+                plugin.getLogger().log( Level.INFO, "Successfully converted the players database..." );
             }
 
             // Adds raceMaster arg to the player table
@@ -120,7 +117,7 @@ class DBWrapperSQLite
             }
             catch ( Exception e )
             {
-                plugin.getLogger().log( Level.INFO, "Converting Player DB (may lag a little wait for completion message)." );
+                plugin.getLogger().log( Level.INFO, "Converting player database..." );
                 mDBCon.setAutoCommit( false );
                 HashMap<UUID, String> dcplayers = new HashMap<>();
                 HashMap<UUID, Integer> ids = new HashMap<>();
@@ -142,7 +139,7 @@ class DBWrapperSQLite
                     e1.printStackTrace();
                 }
                 statement.executeUpdate( "DROP TABLE players" );
-                statement.executeUpdate( "create table players ( id INTEGER PRIMARY KEY, uuid, race, raceMaster );" );
+                statement.executeUpdate( "CREATE TABLE IF NOT EXISTS `" + database + "`. `players` ( `id` INT NOT NULL AUTO_INCREMENT, `uuid` TEXT, `race` TEXT, `raceMaster` TEXT, PRIMARY KEY( `id` ))ENGINE = InnoDB" );
                 for ( UUID uuid : dcplayers.keySet() )
                 {
                     if ( uuid != null )
@@ -156,42 +153,11 @@ class DBWrapperSQLite
                         prep.close();
                     }
                 }
-                plugin.getLogger().log( Level.INFO, "Finished Converting the Players DB." );
+                plugin.getLogger().log( Level.INFO, "Successfully converted the players database..." );
             }
 
-            try
-            {
-                rs = statement.executeQuery( "select * from sqlite_master WHERE name = 'trainers';" );
-                if ( rs.next() )
-                {
-                    plugin.getLogger().log( Level.INFO, "Transferring Trainer DB to citizens  (may lag a little wait for completion message)." );
-
-                    rs = statement.executeQuery( "select * from trainers;" );
-
-                    while ( rs.next() )
-                    {
-                        AbstractNPC npc1;
-                        if ( rs.getString( "type" ).equalsIgnoreCase( "PLAYER" ) )
-                        {
-                            npc1 = ( AbstractNPC ) plugin.getNPCRegistry().createNPC( EntityType.PLAYER, UUID.randomUUID(), Integer.parseInt( rs.getString( "uniqueId" ) ), rs.getString( "name" ) );
-                        }
-                        else
-                        {
-                            npc1 = ( AbstractNPC ) plugin.getNPCRegistry().createNPC( EntityType.valueOf( rs.getString( "type" ) ), UUID.randomUUID(), Integer.parseInt( rs.getString( "uniqueId" ) ), rs.getString( "name" ) );
-                        }
-                        npc1.spawn( new Location( plugin.getServer().getWorld( rs.getString( "world" ) ), rs.getDouble( "x" ), rs.getDouble( "y" ), rs.getDouble( "z" ), rs.getFloat( "yaw" ), rs.getFloat( "pitch" ) ) );
-                        npc1.addTrait( new DwarfTrainerTrait( plugin, Integer.parseInt( rs.getString( "uniqueId" ) ), rs.getInt( "skill" ), rs.getInt( "maxSkill" ), rs.getInt( "minSkill" ) ) );
-                        npc1.setProtected( true );
-                    }
-                }
-                statement.execute( "DROP TABLE trainers" );
-                plugin.getLogger().log( Level.INFO, "Finished Transferring the Trainers DB." );
-            }
-            catch ( Exception e )
-            {
-                // NOOP
-            }
-            rs.close();
+            if ( rs != null )
+                rs.close();
             mDBCon.setAutoCommit( true );
         }
         catch ( Exception e )
@@ -202,23 +168,24 @@ class DBWrapperSQLite
 
     private void buildDB()
     {
+        plugin.getLogger().log( Level.INFO, "Attempting to build database....");
+
         try
         {
             Statement statement = mDBCon.createStatement();
-            ResultSet rs = statement.executeQuery( "select * from sqlite_master WHERE name = 'players';" );
+
+            // Attempt to build database if it cannot be found
+            ResultSet rs = statement.executeQuery( "SELECT SCHEMA_NAME FROM INFORMATION_SCHEMA.SCHEMATA WHERE SCHEMA_NAME = '" + database + "'" );
             if ( !rs.next() )
             {
-                statement.executeUpdate( "create table players ( id INTEGER PRIMARY KEY, uuid, race, raceMaster );" );
+                statement.executeQuery( "CREATE DATABASE IF NOT EXISTS '" + database + "'" );
             }
-            rs.close();
 
-            rs = statement.executeQuery( "select * from sqlite_master WHERE name = 'skills';" );
-            if ( !rs.next() )
-            {
-                statement.executeUpdate( "CREATE TABLE 'skills' " + "  ( " + "    'player' INT, " + "    'id' int, " + "    'level' INT DEFAULT 0, " + "    'deposit1' INT DEFAULT 0, " + "    'deposit2' INT DEFAULT 0, " + "    'deposit3' INT DEFAULT 0, " + "    PRIMARY KEY ('player','id') " + "  );" );
-            }
-            rs.close();
+            // Attempt to build tables if they cannot be found
+            statement.executeUpdate( "CREATE TABLE IF NOT EXISTS `" + database + "`.`players` (`id` INT NOT NULL AUTO_INCREMENT, `uuid` TEXT,`race` TEXT,`raceMaster` TEXT,PRIMARY KEY (`id`)) ENGINE = InnoDB;" );
+            statement.executeUpdate( "CREATE TABLE IF NOT EXISTS `" + database + "`.`skills` ( `player` INT NOT NULL , `id` INT NOT NULL , `level` INT NOT NULL DEFAULT '0' , `deposit1` INT NOT NULL DEFAULT '0' , `deposit2` INT NOT NULL DEFAULT '0' , `deposit3` INT NOT NULL DEFAULT '0' , PRIMARY KEY (`player`, `id`)) ENGINE = InnoDB;" );
 
+            rs.close();
         }
         catch ( SQLException e )
         {
